refactor: вынести утилиты в отдельный модуль
