+++ backend/blogapp/modules/comments/routes.py
@@ -17,12 +17,9 @@ from .models import (
 )
 from ..articles.models import ArticleDocument
 from ..users.models import UserDocument
-from ...core.database.extended_document import (
-    delete_document_by_id,
-    update_document_by_id,
-)
 from ...core.security.roles import RolesEnum
 from ...core.security.utilities import RoleChecker
+from ...utils.document_utils import update_document_by_id, delete_document_by_id
 
 router = APIRouter(prefix="/comments")
